requirements.txt
Flask==2.0.0
Werkzeug>=2.0,<3
stripe>=5.0
python-dotenv==0.19.0